MemoryManager: Try search secrets by looped ASM call instead of behav ID
Trying a new SearchObject-like function. The idea is to take the pointer
to behav script of the current object (already retrieved), but read into
it through the script commands, find any 0C calls in a 08-09 loop, and
if THAT matches the specified function address (pass in a reds/secrets
function), consider the object a collectible.

A direct example that this tries to detect is Shining Stars 4 secrets -
custom behav script, but using the vanilla bhv_hidden_star_trigger_loop.