ROMManager: Search objects by looped ASM call instead of behav ID
Adds support for Shining Stars 4's secrets.

Same algorithm as in MemoryManager (commit 72fb886 and fixes up to
ed87a71), but based on behav scripts' ROM addresses instead of live
process addresses. Those are deduced from level scripts' loading of
segment 13. On the way it is assumed that the segment 15 start address
in ROM is constant.

The implementation in GetAmountOfObjectsInternal has hacky parts (filling
the 3-byte behavior array to 4 bytes, reliance on segment 13, reliance on
indices for read bytes). Future commits are welcome to clean up.